refactor: use custom hook to simplify component
